把图片数据集的逻辑整合到controller里面
