+++ packages/web/i18n/en/common.json
@@ -184,7 +184,6 @@
   "comfirm_leave_page": "Confirm to Leave This Page?",
   "comfirn_create": "Confirm Creation",
   "commercial_function_tip": "Please Upgrade to the Commercial Version to Use This Feature: https://doc.fastgpt.cn/docs/commercial/intro/",
-  "common.Confirm Update": "Confirm update",
   "common.Create Failed": "Failed to create a collection",
   "comon.Continue_Adding": "Continue Adding",
   "compliance.chat": "The content is generated by third-party AI and cannot be guaranteed to be true and accurate. It is for reference only.",
