+++ packages/web/i18n/zh-CN/common.json
@@ -184,7 +184,6 @@
   "comfirm_leave_page": "确认离开该页面？",
   "comfirn_create": "确认创建",
   "commercial_function_tip": "请升级商业版后使用该功能：https://doc.fastgpt.cn/docs/commercial/intro/",
-  "common.Confirm Update": "确认更新",
   "common.Create Failed": "创建集合失败",
   "common.error.unKnow": "未知问题",
   "comon.Continue_Adding": "继续添加",
@@ -558,7 +557,7 @@
   "core.dataset.training.Agent queue": "QA 训练排队",
   "core.dataset.training.Auto mode": "补充索引",
   "core.dataset.training.Auto mode Tip": "通过子索引以及调用模型生成相关问题与摘要，来增加数据块的语义丰富度，更利于检索。需要消耗更多的存储空间和增加 AI 调用次数。",
-  "core.dataset.training.Chunk mode": "分块存储",
+  "core.dataset.training.Chunk mode": "直接分块",
   "core.dataset.training.Full": "预计 20 分钟以上",
   "core.dataset.training.Leisure": "空闲",
   "core.dataset.training.QA mode": "问答对提取",
